fix: fix kl-dom mixtral and add script

+++ mcsmoe/msmoe-merging-mixtral.py
@@ -41,7 +41,7 @@ def evaluate_mcsmoe(
     tokenizer.pad_token_id = tokenizer.eos_token_id
     model = MixtralForCausalLM.from_pretrained(
         model_name,
-        torch_dtype=torch.bfloat16, device_map="auto"
+        torch_dtype=torch.float16, device_map="auto"
     )
     model.eval()
 
@@ -104,22 +104,6 @@ def evaluate_mcsmoe(
         raise ValueError(
             f"Accepted dominant methods are `random`, `frequency` and `knowledge`, but you input {dominant}"
         )
-
-    # Freq-merge
-    # model = merge_by_groups_with_usage_weighted(
-    #     model, grouper=grouper, merging_layers=list(range(0, model.config.num_hidden_layers))
-    # )
-
-    # ZipIt merge
-    # model = merge_by_groups_within_and_across_models(
-    #     mixtral_model=model,
-    #     grouper=grouper,
-    #     dataloader=dataloader_for_merging,
-    #     mode=mode,
-    #     partition=partition,
-    #     dominant_alone=False,
-    #     usage_weighted=False
-    # )
 
     # dom_experts = grouper.core_experts
